Added role and admin rights for comments for methods deleteComment and updateComment

+++ src/main/java/com/skypro/adsonline/service/impl/CommentServiceImpl.java
@@ -2,6 +2,7 @@ package com.skypro.adsonline.service.impl;
 
 import com.skypro.adsonline.dto.Comment;
 import com.skypro.adsonline.dto.ResponseWrapperComment;
+import com.skypro.adsonline.dto.Role;
 import com.skypro.adsonline.exception.AdNotFoundException;
 import com.skypro.adsonline.exception.CommentNotFoundException;
 import com.skypro.adsonline.model.AdEntity;
@@ -11,12 +12,14 @@ import com.skypro.adsonline.repository.CommentRepository;
 import com.skypro.adsonline.security.SecurityUser;
 import com.skypro.adsonline.service.CommentService;
 import com.skypro.adsonline.utils.CommentMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
-import static com.skypro.adsonline.constant.ErrorMessage.AD_NOT_FOUND_MSG;
-import static com.skypro.adsonline.constant.ErrorMessage.COMMENT_NOT_FOUND_MSG;
+import static com.skypro.adsonline.constant.ErrorMessage.*;
 
 @Service
 public class CommentServiceImpl implements CommentService {
@@ -55,20 +58,39 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public boolean deleteComment(Integer adId, Integer commentId) {
+    public boolean deleteComment(Integer adId, Integer commentId, SecurityUser currentUser) {
+        checkPermission(commentId, currentUser);
         CommentEntity comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentNotFoundException(COMMENT_NOT_FOUND_MSG.formatted(commentId)));
         commentRepository.delete(comment);
         return true;
     }
 
     @Override
-    public Comment updateComment(Integer adId, Integer commentId, Comment comment) {
+    public Comment updateComment(Integer adId, Integer commentId, Comment comment, SecurityUser currentUser) {
         AdEntity ad = adRepository.findById(adId).
                 orElseThrow(() -> new AdNotFoundException(AD_NOT_FOUND_MSG.formatted(adId)));
+        checkPermission(commentId, currentUser);
         CommentEntity foundComment = commentRepository.findById(commentId).
                 orElseThrow(() -> new CommentNotFoundException(COMMENT_NOT_FOUND_MSG.formatted(commentId)));
         foundComment.setText(comment.getText());
         commentRepository.save(foundComment);
         return commentMapper.mapToCommentDto(foundComment);
+    }
+    /**
+     * Checks if the author of the comment is the same as logged-in user.
+     * If it is not and the user is not an ADMIN then throw an exception.
+     * @param commentId comment id
+     * @param currentUser logged-in user
+     */
+    private void checkPermission(Integer commentId, SecurityUser currentUser) {
+        Integer authorId = commentRepository
+                .findById(commentId)
+                .orElseThrow(() -> new CommentNotFoundException(COMMENT_NOT_FOUND_MSG.formatted(commentId)))
+                .getAuthor()
+                .getId();
+        if (!authorId.equals(currentUser.getUser().getId())
+                && !currentUser.getAuthorities().contains(new SimpleGrantedAuthority(Role.ADMIN.name()))) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, ACCESS_DENIED_MSG);
+        }
     }
 }
